refract: pst functions
